Implement string variants of RsValue and SharedRsValue
Define types `ConstString`, `OwnedRedisString`, `OwnedRmAllocString`,
`RedisStringRef`, and `RsValueString`, which are the different kinds
of string `RsValue` and `SharedRsValue` will have to hold.

Add variants to `RsValueInteral` for these kinds of strings, and
add corresponding constructors to the `Value` trait.

Update `redis_mock` so that it provides access to `RSDummyContext`,
pointing it to `redis_module::Context::dummy()`

Implement the various constructors for string variants
of `RsValue` and `SharedRsValue` that are in use by the
C codebase` in `value_ffi`.